feat: update landing page layout and remove unused video components
