+++ app/investor/(auth)/page.tsx
@@ -25,7 +25,7 @@ export default function InvestorIntroPage() {
     <div className="min-h-screen flex flex-col items-center justify-center p-4">
       <div className="max-w-3xl w-full space-y-8 text-center">
         {/* Video Container */}
-        <div className="relative w-full aspect-video rounded-lg overflow-hidden bg-muted">
+        {/* <div className="relative w-full aspect-video rounded-lg overflow-hidden bg-muted">
           <video
             className="w-full h-full object-cover"
             src="/videos/intro.mp4"
@@ -33,7 +33,7 @@ export default function InvestorIntroPage() {
           >
             Your browser does not support the video tag.
           </video>
-        </div>
+        </div> */}
 
         {/* Text Content */}
         <h1 className="text-2xl font-bold tracking-tight text-white">
