+++ app/preregister/page.tsx
@@ -1,0 +1,256 @@
+"use client"
+import { useState } from "react"
+import { Button } from "@/components/ui/button"
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog"
+import { Input } from "@/components/ui/input"
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+} from "@/components/ui/form"
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useForm } from "react-hook-form"
+import * as z from "zod"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
+
+// Add this constant for countries
+const COUNTRIES = [
+  { value: 'af', label: 'Afghanistan' },
+  { value: 'al', label: 'Albania' },
+  { value: 'dz', label: 'Algeria' },
+  { value: 'ar', label: 'Argentina' },
+  { value: 'au', label: 'Australia' },
+  { value: 'at', label: 'Austria' },
+  { value: 'bh', label: 'Bahrain' },
+  { value: 'bd', label: 'Bangladesh' },
+  { value: 'be', label: 'Belgium' },
+  { value: 'br', label: 'Brazil' },
+  { value: 'bg', label: 'Bulgaria' },
+  { value: 'kh', label: 'Cambodia' },
+  { value: 'ca', label: 'Canada' },
+  { value: 'cl', label: 'Chile' },
+  { value: 'cn', label: 'China' },
+  { value: 'co', label: 'Colombia' },
+  { value: 'hr', label: 'Croatia' },
+  { value: 'cy', label: 'Cyprus' },
+  { value: 'cz', label: 'Czech Republic' },
+  { value: 'dk', label: 'Denmark' },
+  { value: 'eg', label: 'Egypt' },
+  { value: 'ee', label: 'Estonia' },
+  { value: 'fi', label: 'Finland' },
+  { value: 'fr', label: 'France' },
+  { value: 'de', label: 'Germany' },
+  { value: 'gr', label: 'Greece' },
+  { value: 'hk', label: 'Hong Kong' },
+  { value: 'hu', label: 'Hungary' },
+  { value: 'is', label: 'Iceland' },
+  { value: 'in', label: 'India' },
+  { value: 'id', label: 'Indonesia' },
+  { value: 'ir', label: 'Iran' },
+  { value: 'iq', label: 'Iraq' },
+  { value: 'ie', label: 'Ireland' },
+  { value: 'il', label: 'Israel' },
+  { value: 'it', label: 'Italy' },
+  { value: 'jp', label: 'Japan' },
+  { value: 'jo', label: 'Jordan' },
+  { value: 'kz', label: 'Kazakhstan' },
+  { value: 'ke', label: 'Kenya' },
+  { value: 'kr', label: 'South Korea' },
+  { value: 'kw', label: 'Kuwait' },
+  { value: 'lb', label: 'Lebanon' },
+  { value: 'my', label: 'Malaysia' },
+  { value: 'mx', label: 'Mexico' },
+  { value: 'ma', label: 'Morocco' },
+  { value: 'nl', label: 'Netherlands' },
+  { value: 'nz', label: 'New Zealand' },
+  { value: 'ng', label: 'Nigeria' },
+  { value: 'no', label: 'Norway' },
+  { value: 'om', label: 'Oman' },
+  { value: 'pk', label: 'Pakistan' },
+  { value: 'pa', label: 'Panama' },
+  { value: 'pe', label: 'Peru' },
+  { value: 'ph', label: 'Philippines' },
+  { value: 'pl', label: 'Poland' },
+  { value: 'pt', label: 'Portugal' },
+  { value: 'qa', label: 'Qatar' },
+  { value: 'ro', label: 'Romania' },
+  { value: 'ru', label: 'Russia' },
+  { value: 'sa', label: 'Saudi Arabia' },
+  { value: 'sg', label: 'Singapore' },
+  { value: 'za', label: 'South Africa' },
+  { value: 'es', label: 'Spain' },
+  { value: 'se', label: 'Sweden' },
+  { value: 'ch', label: 'Switzerland' },
+  { value: 'tw', label: 'Taiwan' },
+  { value: 'th', label: 'Thailand' },
+  { value: 'tr', label: 'Turkey' },
+  { value: 'ae', label: 'United Arab Emirates' },
+  { value: 'gb', label: 'United Kingdom' },
+  { value: 'us', label: 'United States' },
+  { value: 'vn', label: 'Vietnam' },
+] as const
+
+// Update the form schema to include investorsDaftar
+const formSchema = z.object({
+  firstName: z.string().min(2, "First name is required"),
+  lastName: z.string().min(2, "Last name is required"),
+  email: z.string().email("Invalid email address"),
+  country: z.string().min(1, "Country is required"),
+  investorsDaftar: z.string().min(2, "Investor's Daftar is required"),
+})
+
+export default function LandingPage() {
+  const [showForm, setShowForm] = useState(false)
+  const [registered, setRegistered] = useState(false)
+
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      firstName: "",
+      lastName: "",
+      email: "",
+      country: "",
+      investorsDaftar: "",
+    },
+  })
+
+  function onSubmit(values: z.infer<typeof formSchema>) {
+    // Here you would typically send the data to your backend
+    console.log(values)
+    setRegistered(true)
+  }
+
+  return (
+    <div className="flex flex-col items-start justify-center h-[60%] px-4">
+      <div className="w-full">
+        {/* Header */}
+        <div>
+          <h2 className="text-2xl font-medium text-muted-foreground mt-[4rem] mb-[16rem]">
+            Daftar OS Technology
+          </h2>
+          <h1 className="text-4xl font-bold tracking-tight mb-8">
+          Helping investors to scout the next meaningful startup. 
+          </h1>
+          <Button 
+            size="lg" 
+            className="text-lg px-8"
+            onClick={() => setShowForm(true)}
+          >
+            Register Me for Early Access
+          </Button>
+        </div>
+
+        {/* Registration Form Dialog */}
+        <Dialog open={showForm} onOpenChange={setShowForm}>
+          <DialogContent className="sm:max-w-[425px]">
+            <DialogHeader>
+              <DialogTitle>
+                {registered ? "Registration Successful!" : "Pre-registration Form"}
+              </DialogTitle>
+            </DialogHeader>
+
+            {registered ? (
+              <div className="text-center py-6 space-y-4">
+                <p className="text-green-600 font-medium">You're in!</p>
+                <p className="text-muted-foreground">
+                  Thanks for signing up for early access. You'll receive an email invite as soon as the software is live.
+                </p>
+              </div>
+            ) : (
+              <Form {...form}>
+                <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+                  <FormField
+                    control={form.control}
+                    name="firstName"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>First Name</FormLabel>
+                        <FormControl>
+                          <Input placeholder="John" {...field} />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="lastName"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Last Name</FormLabel>
+                        <FormControl>
+                          <Input placeholder="Doe" {...field} />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="email"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Email</FormLabel>
+                        <FormControl>
+                          <Input placeholder="john@example.com" type="email" {...field} />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="country"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Country</FormLabel>
+                        <Select onValueChange={field.onChange} defaultValue={field.value}>
+                          <FormControl>
+                            <SelectTrigger>
+                              <SelectValue placeholder="Select your country" />
+                            </SelectTrigger>
+                          </FormControl>
+                          <SelectContent className="max-h-[300px]">
+                            {COUNTRIES.map((country) => (
+                              <SelectItem key={country.value} value={country.value}>
+                                {country.label}
+                              </SelectItem>
+                            ))}
+                          </SelectContent>
+                        </Select>
+                      </FormItem>
+                    )}
+                  />
+
+                  <FormField
+                    control={form.control}
+                    name="investorsDaftar"
+                    render={({ field }) => (
+                      <FormItem>
+                        <FormLabel>Investor's Daftar</FormLabel>
+                        <FormControl>
+                          <Input placeholder="Enter your investor's daftar" {...field} />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+
+                  <Button type="submit" className="w-full">
+                    Register
+                  </Button>
+                </form>
+              </Form>
+            )}
+          </DialogContent>
+        </Dialog>
+      </div>
+    </div>
+  )
+}
